Got PIDController to work, and put it on all testing files
Got PIDController to work, and put it on all testing files. Trapezoildal profiling is not needed anymore, but keep it for learning or last minute purposes. SlewRateLimiter is yet to be implemented but is going to be with PIDController to make it faster and make it so that it doesn't stop super fast.